Issue #39 Change order type of annotations from int do touble

+++ src/main/java/org/jboss/qa/jenkins/test/executor/phase/cleanup/CleanUp.java
@@ -39,5 +39,5 @@ public @interface CleanUp {
 	boolean cleanWorkspace() default false;
 
 	@Order
-	int order() default 0;
+	double order() default 0;
 }
